BagUtilsTest Complete
Added tests for every method (except the private constructor which is
commented as such: "Instantiation of BagUtils is not intended or
required."
Tested null inputs, as the intended behavior on null input is mentioned
in each method's documentation. This can be considered using equivalence
classes (Black box testing technique), as all non-null inputs are
treated the same by these methods, while null is treated different.